Refactor to templates
