+++ src/pages/contact.tsx
@@ -1,13 +1,13 @@
 import React from "react"
-import { Box } from "rebass/styled-components"
 import Layout from "src/components/layout"
 import { SEO } from "src/components/seo"
+import { ContactTemplate } from "src/templates/ContactTemplate"
 
 export default function ContactPage() {
   return (
     <Layout>
       <SEO title="Contact" />
-      <Box>Contact</Box>
+      <ContactTemplate />
     </Layout>
   )
 }
