+++ src/pages/listen-here.tsx
@@ -1,340 +1,13 @@
 import React from "react"
 import Layout from "src/components/layout"
-import { Box, Flex, Text } from "rebass/styled-components"
 import { SEO } from "src/components/seo"
+import { ListenHereTemplate } from "src/templates/ListenHereTemplate"
 
 const ListenHerePage = _props => {
   return (
     <Layout>
       <SEO title="Listen Here" />
-      <Box style={{ textAlign: "center" }}>
-        <h1>Toolshed SoundLab Productions - George Rezendes</h1>
-        <h4>
-          These are some of the tunes I’ve recorded, mixed and / or produced
-          since 1998. Enjoy!
-        </h4>
-      </Box>
-      <Box>
-        <Box>
-          <Box>
-            <iframe
-              width="100%"
-              height="166"
-              scrolling="no"
-              frameBorder="no"
-              src="https://w.soundcloud.com/player/?url=https%3A//api.soundcloud.com/tracks/324335863&amp;color=ff9900&amp;auto_play=false&amp;hide_related=false&amp;show_comments=true&amp;show_user=true&amp;show_reposts=false"
-            ></iframe>
-          </Box>
-          <Box>
-            <Box>
-              <Text fontWeight="bold" fontSize={1}>
-                Alanna Dailey - I’ve Got To Use My Imagination{" "}
-              </Text>
-            </Box>
-            <Box>
-              <Text fontSize={1} lineHeight={1.5}>
-                Alanna is a talented singer raised here in Port Townsend. She
-                started singing classic jazz and old blues at an early age and
-                was mentored by some of our town's finest musicians including
-                Pete Toyne and Jenny Davis. Alanna finished tracking all of her
-                parts for an eclectic album we recorded in the summer of 2015
-                before she moved to France. This is the first track completed
-                and has a video to go with it. https://youtu.be/lfhoIPpFIbk
-              </Text>
-            </Box>
-          </Box>
-        </Box>
-
-        <Box my={4}>
-          <hr />
-        </Box>
-
-        <Box>
-          <Box>
-            <iframe
-              width="100%"
-              height="166"
-              scrolling="no"
-              frameBorder="no"
-              src="https://w.soundcloud.com/player/?url=https%3A//api.soundcloud.com/tracks/228159938&amp;color=bd7912&amp;auto_play=false&amp;hide_related=false&amp;show_comments=true&amp;show_user=true&amp;show_reposts=false"
-            ></iframe>
-          </Box>
-          <Box>
-            <Box>
-              <Text fontWeight="bold" fontSize={1}>
-                Dudu Maia & Douglas Lora - Choro Pra Je #2
-              </Text>
-            </Box>
-            <Box>
-              <Text fontSize={1} lineHeight={1.5}>
-                Dudu & Douglas spent a few days at the studio playing with mics
-                and gear and talking shop. Dudu has a very nice and active home
-                studio in Brazil where he makes amazing recordings and videos of
-                this remarkable music. Casa Do Som -
-                www.dudumaia.com/casa-do-som) They are both dedicated &
-                masterful musicians and it was a great pleasure to record them.
-              </Text>
-            </Box>
-          </Box>
-        </Box>
-
-        <Box my={4}>
-          <hr />
-        </Box>
-
-        <Box>
-          <Box>
-            <iframe
-              width="100%"
-              height="166"
-              scrolling="no"
-              frameBorder="no"
-              src="https://w.soundcloud.com/player/?url=https%3A//api.soundcloud.com/tracks/228582085&amp;color=af7011&amp;auto_play=false&amp;hide_related=false&amp;show_comments=true&amp;show_user=true&amp;show_reposts=false"
-            ></iframe>
-          </Box>
-          <Box>
-            <Box>
-              <Text fontWeight="bold" fontSize={1}>
-                Rattletrap Ruckus - Easy Winner
-              </Text>
-            </Box>
-            <Box>
-              <Text fontSize={1} lineHeight={1.5}>
-                Rattletrap Ruckus - Really cool people playing incredibly cool
-                music - I'm at a loss for words. Listen up and follow the yellow
-                bit link to the wild musical work of the Ruckus ones!
-                http://rattletrapruckus.com
-              </Text>
-            </Box>
-          </Box>
-        </Box>
-
-        <Box my={4}>
-          <hr />
-        </Box>
-
-        <Box>
-          <Box>
-            <iframe
-              width="100%"
-              height="166"
-              scrolling="no"
-              frameBorder="no"
-              src="https://w.soundcloud.com/player/?url=https%3A//api.soundcloud.com/tracks/228580824&amp;color=b67615&amp;auto_play=false&amp;hide_related=false&amp;show_comments=true&amp;show_user=true&amp;show_reposts=false"
-            ></iframe>
-          </Box>
-          <Box>
-            <Box>
-              <Text fontWeight="bold" fontSize={1}>
-                Ben Hunter & Joe Seamons - Jazz Fiddler
-              </Text>
-            </Box>
-            <Box>
-              <Text fontSize={1} lineHeight={1.5}>
-                Ben Hunter and Joe Seamons are always a pleasure to have at the
-                studio. They stay for a few days and make such glorious
-                fun-filled music with the perfect balance of care and
-                recklessness. This is from their first recording here in 2014
-                called "Take Yo Time". Here's a link to thee fine fellas
-                @benjoemusic Enjoy!
-              </Text>
-            </Box>
-          </Box>
-        </Box>
-
-        <Box my={4}>
-          <hr />
-        </Box>
-
-        <Box>
-          <Box>
-            <iframe
-              width="100%"
-              height="166"
-              scrolling="no"
-              frameBorder="no"
-              src="https://w.soundcloud.com/player/?url=https%3A//api.soundcloud.com/tracks/228578499&amp;color=a8690c&amp;auto_play=false&amp;hide_related=false&amp;show_comments=true&amp;show_user=true&amp;show_reposts=false"
-            ></iframe>
-          </Box>
-          <Box>
-            <Box>
-              <Text fontWeight="bold" fontSize={1}>
-                Tyva Kyzy - Ayak Shaiym (My Bowl Of Tea)
-              </Text>
-            </Box>
-            <Box>
-              <Text fontSize={1} lineHeight={1.5}>
-                These five women flew from Tuva to Port Townsend to make this
-                recording and then do a west coast tour. They spoke little
-                English, I spoke no Tuvan and we just set up and worked for 5
-                days. I mixed and mastered while they were on the tour, taking
-                orders for the CD along the way. At the end of the tour we
-                delivered the finished CD's. It was a delightful experience.
-              </Text>
-            </Box>
-          </Box>
-        </Box>
-
-        <Box my={4}>
-          <hr />
-        </Box>
-
-        <Box>
-          <Box>
-            <iframe
-              width="100%"
-              height="166"
-              scrolling="no"
-              frameBorder="no"
-              src="https://w.soundcloud.com/player/?url=https%3A//api.soundcloud.com/tracks/229648435&amp;color=af7115&amp;auto_play=false&amp;hide_related=false&amp;show_comments=true&amp;show_user=true&amp;show_reposts=false"
-            ></iframe>
-          </Box>
-          <Box>
-            <Box>
-              <Text fontWeight="bold" fontSize={1}>
-                The Blackberry Bushes - Cluck Old Hen
-              </Text>
-            </Box>
-            <Box>
-              <Text fontSize={1} lineHeight={1.5}>
-                The BB’s came by the studio to film a video for their tour
-                promotion. This was the highlight tune. They played around a
-                single main mic with 2 side fill mics for solos and got a great
-                honest 'in the room' sound. The video can be seen here:
-                youtu.be/9VdG1oicOBA And find out more about them here:
-                www.theblackberrybushes.com
-              </Text>
-            </Box>
-          </Box>
-        </Box>
-
-        <Box my={4}>
-          <hr />
-        </Box>
-
-        <Box>
-          <Box>
-            <iframe
-              width="100%"
-              height="166"
-              scrolling="no"
-              frameBorder="no"
-              src="https://w.soundcloud.com/player/?url=https%3A//api.soundcloud.com/tracks/228580260&amp;color=0066cc&amp;auto_play=false&amp;hide_related=false&amp;show_comments=true&amp;show_user=true&amp;show_reposts=false"
-            ></iframe>
-          </Box>
-        </Box>
-
-        <Flex flexWrap="wrap">
-          <Flex width="33%" flexDirection="column" alignItems="center">
-            <Box>
-              <img src="images/listen-here/dudu-mala.jpg" />
-            </Box>
-            <Text fontSize={1} lineHeight={1.5} textAlign="center">
-              Dudu Maia & Douglas Loura Vibrações
-            </Text>
-          </Flex>
-
-          <Flex width="33%" flexDirection="column" alignItems="center">
-            <Box>
-              <img src="images/listen-here/stubby.jpg" />
-            </Box>
-            <Text fontSize={1} lineHeight={1.5} textAlign="center">
-              Stubby & the Electric Blanket on and on and on
-            </Text>
-          </Flex>
-
-          <Flex width="33%" flexDirection="column" alignItems="center">
-            <Box>
-              <img src="images/listen-here/sam.jpg" />
-            </Box>
-            <Text fontSize={1} lineHeight={1.5} textAlign="center">
-              Sam Doores Farmland Blues
-            </Text>
-          </Flex>
-
-          <Flex width="33%" flexDirection="column" alignItems="center">
-            <Box>
-              <img src="images/listen-here/kate.jpg" />
-            </Box>
-            <Text fontSize={1} lineHeight={1.5} textAlign="center">
-              Kate Long & Robin Kessinger Red Rockin’ Chair Live at Toolshed
-              Soundlab
-            </Text>
-          </Flex>
-
-          <Flex width="33%" flexDirection="column" alignItems="center">
-            <Box>
-              <img src="images/listen-here/gallus.jpg" />
-            </Box>
-            <Text fontSize={1} lineHeight={1.5} textAlign="center">
-              Gallus Bros. Instrumental
-            </Text>
-          </Flex>
-
-          <Flex width="33%" flexDirection="column" alignItems="center">
-            <Box>
-              <img src="images/listen-here/paul.jpg" />
-            </Box>
-            <Text fontSize={1} lineHeight={1.5} textAlign="center">
-              Paul Chasman / Last of Us Soundtrack
-            </Text>
-          </Flex>
-
-          <Flex width="33%" flexDirection="column" alignItems="center">
-            <Box>
-              <img src="images/listen-here/usana.jpg" />
-            </Box>
-            <Text fontSize={1} lineHeight={1.5} textAlign="center">
-              Usana Great Day
-            </Text>
-          </Flex>
-
-          <Flex width="33%" flexDirection="column" alignItems="center">
-            <Box>
-              <img src="images/listen-here/crow.jpg" />
-            </Box>
-            <Text fontSize={1} lineHeight={1.5} textAlign="center">
-              Crow Quill Night Owls Wake Up Sinners
-            </Text>
-          </Flex>
-
-          <Flex width="33%" flexDirection="column" alignItems="center">
-            <Box>
-              <img src="images/listen-here/gallus-2.jpg" />
-            </Box>
-            <Text fontSize={1} lineHeight={1.5} textAlign="center">
-              Gallus Bros, mighty good leader
-            </Text>
-          </Flex>
-
-          <Flex width="33%" flexDirection="column" alignItems="center">
-            <Box>
-              <img src="images/listen-here/inkwell.jpg" />
-            </Box>
-            <Text fontSize={1} lineHeight={1.5} textAlign="center">
-              Inkwell Rhythm Makers Oh Mama
-            </Text>
-          </Flex>
-
-          <Flex width="33%" flexDirection="column" alignItems="center">
-            <Box>
-              <img src="images/listen-here/below.jpg" />
-            </Box>
-            <Text fontSize={1} lineHeight={1.5} textAlign="center">
-              Below the Salt Alliswinigitator
-            </Text>
-          </Flex>
-
-          <Flex width="33%" flexDirection="column" alignItems="center">
-            <Box>
-              <img src="images/listen-here/black.jpg" />
-            </Box>
-            <Text fontSize={1} lineHeight={1.5} textAlign="center">
-              Black Resonators So Nice
-            </Text>
-          </Flex>
-        </Flex>
-      </Box>
+      <ListenHereTemplate />
     </Layout>
   )
 }
