+++ src/templates/ClassesWorkshopsTemplate.tsx
