+++ src/templates/ContactTemplate.tsx
@@ -1,0 +1,10 @@
+import React from "react"
+import { Box } from "rebass/styled-components"
+
+export const ContactTemplate = _props => {
+  return (
+    <Box>
+      <Box>Contact</Box>
+    </Box>
+  )
+}
